backend/requirements.dev.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
PyYAML==6.0.2
httpx==0.27.2
langchain==0.2.7
langchain-community==0.2.7
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0
pypdf==4.3.1
sqlalchemy==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
